Extract heatpump runtime logic
