+++ custom_components/enphase_ev/const.py
@@ -50,6 +50,14 @@ DEFAULT_API_TIMEOUT = 15
 DEFAULT_SESSION_HISTORY_INTERVAL_MIN = 10
 DEFAULT_NOMINAL_VOLTAGE = 230
 BATTERY_PROFILE_PENDING_TIMEOUT_S = 900.0
+HEMS_SUPPORT_PREFLIGHT_CACHE_TTL = 15.0
+HEATPUMP_RUNTIME_DIAGNOSTICS_CACHE_TTL = 60.0
+HEATPUMP_RUNTIME_STATE_CACHE_TTL = 15.0
+HEATPUMP_RUNTIME_STATE_FAILURE_BACKOFF_S = 900.0
+HEATPUMP_DAILY_CONSUMPTION_CACHE_TTL = 300.0
+HEATPUMP_DAILY_CONSUMPTION_FAILURE_BACKOFF_S = 900.0
+HEATPUMP_POWER_CACHE_TTL = 15.0
+HEATPUMP_POWER_FAILURE_BACKOFF_S = 900.0
 BATTERY_PROFILE_LABELS = {
     "self-consumption": "Self-Consumption",
     "cost_savings": "Savings",
